chore: setup css-modules with scss
