+++ src/components/HomePage/styles.module.scss
@@ -1,0 +1,3 @@
+.root {
+  outline: 1px solid green;
+}
