+++ src/layouts/index.js
@@ -14,17 +14,8 @@ const Layout = ({ children, data }) => (
         { name: 'keywords', content: 'sample, something' },
       ]}
     />
-    <Header siteTitle={data.site.siteMetadata.title} />
-    <div
-      style={{
-        margin: '0 auto',
-        maxWidth: 960,
-        padding: '0px 1.0875rem 1.45rem',
-        paddingTop: 0,
-      }}
-    >
-      {children()}
-    </div>
+    {/* <Header siteTitle={data.site.siteMetadata.title} /> */}
+    <div>{children()}</div>
   </div>
 )
